+++ src/main/java/com/hixon/financialApp/view/spreadsheetXml/SpreadsheetXmlForecastView.java
@@ -327,7 +327,8 @@ public class SpreadsheetXmlForecastView extends AbstractForecastView {
    }
 
    @Override
-   public void renderForecastTransaction(ForecastTransaction forecastTransaction, int credit, int debit) throws EntityException,
+   public void renderForecastTransaction(ForecastTransaction forecastTransaction, double credit, double debit)
+           throws EntityException,
            SQLException, ForecastException, BudgetException {
 
       // Output the forecast transaction and item:
